requirements.txt
Django>1.9.0
pycrest
pgcli
psycopg2
django-redis
pytz
djangorestframework